honor module cache settings when loading templates so caching can be disabled for template testing remove (never used) template caching option (memory vs disk vs database) separate templates caching from DataCache to separate class TemplateCache
