+++ App_LocalResources/ControlPanel.ascx.resx
@@ -402,12 +402,6 @@
   <data name="[RESX:SubscriptionOptions].Text" xml:space="preserve">
     <value>Subscription Options:</value>
   </data>
-  <data name="[RESX:TemplateCacheDisk].Text" xml:space="preserve">
-    <value>File System</value>
-  </data>
-  <data name="[RESX:TemplateCacheMemory].Text" xml:space="preserve">
-    <value>Memory</value>
-  </data>
   <data name="[RESX:TopForums].Text" xml:space="preserve">
     <value>Top Forums</value>
   </data>
