+++ CustomControls/UserControls/ForumView.cs
@@ -27,6 +27,7 @@ using System.Web.UI;
 using System.Web.UI.WebControls;
 using System.Text.RegularExpressions;
 using System.Collections.Generic;
+using System.Reflection;
 
 namespace DotNetNuke.Modules.ActiveForums.Controls
 {
@@ -182,7 +183,6 @@ namespace DotNetNuke.Modules.ActiveForums.Controls
                 SettingsInfo MainSettings = DataCache.MainSettings(ModuleId);
                 string sOutput = string.Empty;
                 string sTemplate;
-                int TemplateCache = MainSettings.TemplateCache;
                 if (UseTemplatePath && TemplatePath != string.Empty)
                 {
                     DisplayTemplate = Utilities.GetFileContent(TemplatePath + "ForumView.htm");
@@ -194,7 +194,7 @@ namespace DotNetNuke.Modules.ActiveForums.Controls
                     DisplayTemplate = Utilities.ParseSpacer(DisplayTemplate);
                 }
 
-                sTemplate = DisplayTemplate == string.Empty ? DataCache.GetCachedTemplate(TemplateCache, ModuleId, "ForumView", ForumTemplateId) : DisplayTemplate;
+                sTemplate = DisplayTemplate == string.Empty ? DotNetNuke.Modules.ActiveForums.TemplateCache.GetCachedTemplate( ModuleId, "ForumView", ForumTemplateId) : DisplayTemplate; 
                 try
                 {
                     sTemplate = ParseControls(sTemplate);
