+++ CustomControls/UserControls/TopicsView.cs
@@ -188,7 +188,7 @@ namespace DotNetNuke.Modules.ActiveForums.Controls
                 }
                 else
                 {
-                    TopicsTemplate = DataCache.GetCachedTemplate(MainSettings.TemplateCache, ModuleId, "TopicsView", defaultTemplateId);
+                    TopicsTemplate = TemplateCache.GetCachedTemplate( ModuleId, "TopicsView", defaultTemplateId);
                 }
                 bool loadComplete = false;
                 if (TopicsTemplate.Contains("[NOTOOLBAR]"))
