+++ class/Settings.cs
@@ -172,12 +172,6 @@ namespace DotNetNuke.Modules.ActiveForums
         {
             get { return MainSettings.GetString(SettingKeys.AllowSubTypes, string.Empty); }
         }
-
-        public int TemplateCache
-        {
-            get { return MainSettings.GetInt(SettingKeys.TemplateCache); }
-        }
-
         public bool MailQueue
         {
             get { return MainSettings.GetBoolean(SettingKeys.MailQueue); }
