+++ config/defaultsetup.config
@@ -3,8 +3,7 @@
   <mainsettings>
     <setting name="THEME" value="_default" />
     <setting name="FORUMTEMPLATEID" value="0" />
-    <setting name="PAGESIZE" value="20" />
-    <setting name="TEMPLATECACHE" value="0" />
+    <setting name="PAGESIZE" value="20" /> 
     <setting name="DATEFORMATSTRING" value="dd MMM yyyy" />
     <setting name="TIMEFORMATSTRING" value="hh:mm tt" />
     <setting name="TIMEZONEOFFSET" value="0" />
